Test.Hspec.Expectations.Contrib

Description

Experimental combinators, that may become part of the main distribution, if they turn out to be useful for a wider audience.

Documentation

module Test.Hspec.Expectations

Predicates

(useful in combination with shouldSatisfy)

isLeft :: Either a b -> Bool

isRight :: Either a b -> Bool