Test.Hspec.Expectations.Contrib

Description

Experimental combinators, that may become part of the main distribution, if they turn out to be useful for a wider audience.

Additional combinators for setting expectations

shouldNotBe :: (Show a, Eq a) => a -> a -> Expectation infix 1

actual `shouldNotBe` notExpected sets the expectation that actual is not equal to notExpected

shouldNotSatisfy :: Show a => a -> (a -> Bool) -> Expectation infix 1

v `shouldNotSatisfy` p sets the expectation that p v is False.

shouldNotReturn :: (Show a, Eq a) => IO a -> a -> Expectation infix 1

action `shouldNotReturn` notExpected sets the expectation that action does not return notExpected.

shouldNotContain :: (Show a, Eq a) => [a] -> [a] -> Expectation infix 1

list `shouldNotContain` sublist sets the expectation that sublist is not contained anywhere in the second.

Predicates

(useful in combination with shouldSatisfy)

isLeft :: Either a b -> Bool

isRight :: Either a b -> Bool